+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # LTI Submissions Plugin
 
+## 2026-03-16 
+- Update plugin incompatibility version to Moodle 502
+- Adds plugin version to custom params
+
 ## 2025-09-17
 - Make HTTP requests using Moodle's curl class to support proxies [#15](https://github.com/Vericus/moodle_lti_submissions/pull/15)